perf: use `uint256`

+++ contracts/libraries/BondPosition.sol
@@ -64,9 +64,9 @@ library BondPositionLibrary {
         Coupon[] memory burnCoupons = new Coupon[](burnCouponsLength);
         mintCouponsLength = 0;
         burnCouponsLength = 0;
-        uint16 farthestExpiredEpochs = newPosition.expiredWith.max(oldPosition.expiredWith).sub(latestExpiredEpoch);
+        uint256 farthestExpiredEpochs = newPosition.expiredWith.max(oldPosition.expiredWith).sub(latestExpiredEpoch);
         unchecked {
-            for (uint16 i = 0; i < farthestExpiredEpochs; ++i) {
+            for (uint256 i = 0; i < farthestExpiredEpochs; ++i) {
                 latestExpiredEpoch = latestExpiredEpoch.add(1); // reuse latestExpiredEpoch as epoch
                 uint256 newAmount = newPosition.expiredWith < latestExpiredEpoch ? 0 : newPosition.amount;
                 uint256 oldAmount = oldPosition.expiredWith < latestExpiredEpoch ? 0 : oldPosition.amount;
